pyproject.toml
[project]
name = "zerocode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Manoj Bajaj", email = "manoj@zetsy.dev" }
]
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.1.24",
    "chromadb>=0.6.3",
    "datasets>=3.3.2",
    "duckduckgo-search>=7.5.0",
    "e2b-code-interpreter>=1.0.5",
    "langchain>=0.3.19",
    "langchain-anthropic>=0.3.8",
    "langchain-chroma>=0.2.2",
    "langchain-cohere>=0.4.2",
    "langchain-elasticsearch>=0.3.2",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.0.2",
    "langchain-mongodb>=0.4.0",
    "langchain-openai>=0.3.6",
    "langchain-pinecone>=0.2.3",
    "langgraph>=0.2.74",
    "langsmith>=0.3.10",
    "litellm>=1.61.13",
    "loguru>=0.7.3",
    "markitdown>=0.0.1a4",
    "matplotlib>=3.10.0",
    "rich>=13.9.4",
    "seaborn>=0.13.2",
    "tavily-python>=0.5.1",
    "typer>=0.15.1",
    "youtube-transcript-api>=0.6.3",
]

[project.scripts]
zerocode = "zerocode:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
lint = [
    "ruff>=0.9.7",
]
